requirements.txt
starlette~=0.37.2
fastapi~=0.111.0
uvicorn~=0.29.0
python-dotenv~=1.0.1
PyJWT~=2.8.0
pyodbc~=5.1.0
pycryptodome~=3.20.0
azure-storage-blob~=12.20.0